requirements.txt
aioredis==1.3.1
asgiref~=3.4.1
astroid==2.0.4
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
autopep8==1.6.0
certifi==2018.10.15
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.0.1
cryptography==36.0.1
daphne==3.0.2
defusedxml==0.5.0
Django~=3.2.8
django-allauth==0.48.0
django-cors-headers==3.11.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Faker==13.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.7
incremental==21.3.0
isort==4.3.4
itypes==1.2.0
Jinja2==3.0.3
joblib==1.1.0
lazy-object-proxy==1.3.1
MarkupSafe==2.1.0
mccabe==0.6.1
msgpack==1.0.3
numpy==1.22.1
oauthlib==2.1.0
pandas==1.3.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.1.1
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python3-openid==3.1.0
pytz==2021.3
requests~=2.27.1
requests-oauthlib==1.0.0
scikit-surprise==1.1.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toml==0.10.2
Twisted==22.1.0
txaio==21.2.1
typed-ast==1.5.2
typing_extensions==4.0.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.24.1
wrapt==1.10.11
zope.interface==5.4.0
pytest~=6.2.5
icalendar==4.0.9
